Supports questions with 4 or less answers; Supports single question.
Now supports any question with 4 or less answers. Also, previously if only one question was present in the questions.js file, it wouldn't allow you to answer the question and end the quiz.

So, a TypeError is now thrown when it tries to pull from a list of zero questions, and it ends the quiz.